backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
django = "^4.2.6"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
psycopg-binary = "^3.1.12"
pydantic = "^2.4.2"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.0"
psycopg = "^3.1.12"
pandas = "^2.1.1"
django-filter = "^23.3"
setuptools = "^68.2.2"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
joblib = "^1.3.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.292"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
